chore(driver): move inline tests to a separate file
The test module was extracted from driver.rs into a dedicated driver_tests.rs file to reduce the size of the main source file and improve maintainability. The module is now included via a path attribute, preserving all existing test behaviour.

Auto-committed-on: dragonfly
